[PL] Rename GroundwaterFlow to SteadyStateDiffusion I.
- Rename
  + process name,
  + process folder, and
  + cmake macros.
